requirements.txt
requests
numpy
flask
Flask-Cors
python-dotenv
groq
deepgram-sdk
astrapy